+++ components/boxes/TagBox/ProfileTagBox.js
@@ -1,17 +1,56 @@
 import React from 'react';
+import {connect} from 'react-redux';
 
-import ProfileTagList from './ProfileTagList';
-import ProfileTagFormContainer from './ProfileTagFormContainer';
+import {
+  addTagToProfile,
+  loadCandidateTags,
+  initializeCandidateTags,
+  deleteTagToProfile
+} from '../../../actions/dispatchers';
+import TagBox from './TagBox';
 
 const ProfileTagBox = React.createClass({
+  propTypes: {
+    profileId: React.PropTypes.number.isRequired
+  },
+
+  getTagFormId() {
+    return `profile-${this.props.profileId}`;
+  },
+
   render() {
     return (
-      <div className="tag-container">
-        <ProfileTagList profileId={this.props.profileId}/>
-        <ProfileTagFormContainer profileId={this.props.profileId}/>
-      </div>
+      <TagBox
+        id={this.props.profileId}
+        addTag={this.props.addTagToProfile}
+        candidates={this.props.candidates}
+        loadCandidateTags={this.props.loadCandidateTags}
+        initializeCandidateTags={this.props.initializeCandidateTags}
+        getTagFormId={this.getTagFormId}
+        userLevel={this.props.userLevel}
+        deletable
+        deleteTag={this.props.deleteTagToProfile}
+        tags={this.props.tags}
+        />
     );
   }
 });
 
-export default ProfileTagBox;
+const mapStateToProps = function (state) {
+  return {
+    userLevel: state.userInfo.userLevel,
+    candidates: state.tag.candidate.byTagFormId,
+    tags: state.tag.attached.profiles
+  };
+};
+
+const mapDispatchToProps = function (dispatch) {
+  return {
+    deleteTagToProfile: (profileId, tagName) => deleteTagToProfile(dispatch, profileId, tagName),
+    addTagToProfile: (profileId, tagName) => addTagToProfile(dispatch, profileId, tagName),
+    initializeCandidateTags: tagFormId => initializeCandidateTags(dispatch, tagFormId),
+    loadCandidateTags: (tagFormId, query) => loadCandidateTags(dispatch, tagFormId, query)
+  };
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(ProfileTagBox);
